# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  api.Comment:
    properties:
      content:
        example: This is a comment
        type: string
      id:
        type: integer
      postId:
        example: 1
        type: integer
      reactionInfo:
        $ref: '#/definitions/api.ReactionInfo'
      user:
        $ref: '#/definitions/api.UserResponseInfo'
    type: object
...
  api.CommentCreateResponse:
    properties:
      comments:
        $ref: '#/definitions/api.Comment'
    type: object
  api.ErrorDetails:
    properties:
      code:
        type: integer
      details:
        items:
          $ref: '#/definitions/api.ValidationError'
        type: array
      message:
        type: string
    type: object
...
  api.Group:
    properties:
      bannerColor:
        example: '#000000'
        type: string
      chatHash:
        type: string
      createdAt:
        type: string
      creatorInfo:
        $ref: '#/definitions/api.UserResponseInfo'
      description:
        example: Test description
        type: string
      events:
        items:
          $ref: '#/definitions/api.GroupEvent'
        type: array
      id:
        type: integer
      memberInfo:
        $ref: '#/definitions/api.MemberInfo'
      members:
        items:
          $ref: '#/definitions/api.GroupMember'
        type: array
      posts:
        items:
          $ref: '#/definitions/api.Post'
        type: array
      title:
        example: Test title
        type: string
    type: object
  api.GroupCreateRequest:
    properties:
      bannerColor:
        example: '#000000'
        type: string
      description:
        example: Test description
        type: string
      title:
        example: Test title
        type: string
    type: object
  api.GroupEvent:
    properties:
      creatorInfo:
        $ref: '#/definitions/api.UserResponseInfo'
      date:
        example: "2021-01-01T00:00:00Z"
        type: string
      description:
        example: Test description
        type: string
      groupId:
        example: 1
        type: integer
      id:
        type: integer
      members:
        items:
          $ref: '#/definitions/api.UserResponseInfo'
        type: array
      title:
        example: Test title
        type: string
      usersGoing:
        type: integer
    type: object
  api.GroupEventCreate:
    properties:
      date:
        example: "2021-01-01T00:00:00Z"
        type: string
      description:
        example: Test description
        type: string
      groupId:
        example: 1
        type: integer
      title:
        example: Test title
        type: string
    type: object
  api.GroupMember:
    properties:
      id:
        type: integer
      nickname:
        type: string
      role:
        type: string
    type: object
  api.GroupRequest:
    properties:
      groupId:
        type: integer
      requestType:
        description: '"inv" or "j_req"'
        example: inv
        type: string
    type: object
  api.GroupResponse:
    properties:
      group:
        $ref: '#/definitions/api.Group'
    type: object
  api.GroupResponseInfo:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  api.MemberInfo:
    properties:
      roles:
        type: string
      totalMembers:
        type: integer
    type: object
  api.Metadata:
    properties:
      timestamp:
        type: string
      version:
        type: string
    type: object
  api.NotificationInfo:
    properties:
      id:
        type: integer
      idRef:
        type: integer
      type:
        type: string
    type: object
  api.NotificationResponse:
    properties:
      notification_info:
        $ref: '#/definitions/api.NotificationInfo'
      requestId:
        example: 1
        type: integer
      status:
        description: '"accepted" or "rejected"'
        example: accepted
        type: string
    type: object
  api.Post:
    properties:
      commentCount:
        type: integer
      content:
        example: Test content
        type: string
      group:
        $ref: '#/definitions/api.GroupResponseInfo'
      id:
        type: integer
      image:
        type: string
      privacy:
        example: public
        type: string
      reactionInfo:
        $ref: '#/definitions/api.ReactionInfo'
      title:
        example: Test title
        type: string
      user:
        $ref: '#/definitions/api.UserResponseInfo'
    type: object
...
  api.ReactionInfo:
    properties:
      dislikeCount:
        type: integer
      likeCount:
        type: integer
      status:
        type: string
    type: object
...
  api.Response:
    properties:
      authenticated:
        type: boolean
      message:
        type: string
      metadata:
        $ref: '#/definitions/api.Metadata'
      status:
        type: string
      user:
        $ref: '#/definitions/api.UserResponseInfo'
    type: object
  api.ResponseID:
    properties:
      id:
        type: integer
    type: object
...
  api.UserResponseInfo:
    properties:
      id:
        type: integer
      nickname:
        type: string
    type: object
...
  api.ValidationError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server for kood-social-network.
  title: kood-social-network API
  version: 0.1.0
paths:
...
  /groups:
    get:
      consumes:
      - application/json
      description: Fetches all groups.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                payload:
                  $ref: '#/definitions/api.CommentCreateResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                eror:
                  $ref: '#/definitions/api.ErrorDetails'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                eror:
                  $ref: '#/definitions/api.ErrorDetails'
              type: object
      summary: Get all groups
      tags:
      - groups
    post:
      consumes:
      - application/json
      description: Creates a new group with the provided title and description.
      parameters:
      - description: Group details
        in: body
        name: GroupCreateRequest
        required: true
        schema:
          $ref: '#/definitions/api.GroupCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                payload:
                  $ref: '#/definitions/api.ResponseID'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                eror:
                  $ref: '#/definitions/api.ErrorDetails'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                eror:
                  $ref: '#/definitions/api.ErrorDetails'
              type: object
      security:
      - BearerAuth: []
      summary: Create a new group
      tags:
      - groups
  /groups/{id}:
    get:
      consumes:
      - application/json
      description: Fetches a group by its ID.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                payload:
                  $ref: '#/definitions/api.GroupResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                eror:
                  $ref: '#/definitions/api.ErrorDetails'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                eror:
                  $ref: '#/definitions/api.ErrorDetails'
              type: object
      security:
      - BearerAuth: []
      summary: Get a group by ID
      tags:
      - groups
  /groups/event:
    post:
      consumes:
      - application/json
      description: Creates a new group event with the provided title, description,
        and date.
      parameters:
      - description: Group event details
        in: body
        name: GroupEventCreateRequest
        required: true
        schema:
          $ref: '#/definitions/api.GroupEventCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                payload:
                  $ref: '#/definitions/api.GroupEvent'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                eror:
                  $ref: '#/definitions/api.ErrorDetails'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                eror:
                  $ref: '#/definitions/api.ErrorDetails'
              type: object
      security:
      - BearerAuth: []
      summary: Create a new group event
      tags:
      - groups
  /groups/request:
    post:
      consumes:
      - application/json
      description: Processes a group request, either accepting or rejecting it.
      parameters:
      - description: Group request details
        in: body
        name: GroupRequest
        required: true
        schema:
          $ref: '#/definitions/api.GroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                eror:
                  $ref: '#/definitions/api.ErrorDetails'
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/api.Response'
            - properties:
                eror:
                  $ref: '#/definitions/api.ErrorDetails'
              type: object
      security:
      - BearerAuth: []
      summary: Process a group request
      tags:
      - groups
  /notification/response:
    post:
      consumes:
      - application/json
      description: Processes the response to a notification, which could involve actions
        such as accepting or rejecting various types of notifications.
      parameters:
      - description: Notification response details
        in: body
        name: NotificationResponse
        required: true
        schema:
          $ref: '#/definitions/api.NotificationResponse'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - BearerAuth: []
      summary: Process a notification response
      tags:
      - notifications
...
swagger: "2.0"